requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0
Flask-Session==0.5.0
itsdangerous==2.1.2
pytz==2023.3.post1